Improvements for customer portal registration.
Use of --autosubscribe during registration as well as subscribing with a
specific pool ID was likely double entitling. Auto-subscribe will handle
getting the system anything it needs assuming it has a RHEL product cert
installed in /etc/pki/product.

Skip attempts to register if the system is already registered. (it's possible
the user already registered to install git and clone katello-deploy repo)

Rather than requiring username/password on the CLI, let subscription-manager
prompt for them if we have to actually do a registration.