Add parameter to allow testing module pull requests.
This adds the ability to specify any number of module/pull request
combinations that are then pulled into a local copy of the installer.
This can be used to more easily test someones change(s) to a given
module or set of modules without having to go through a clone or patch
workflow. See README for examples.