sort scripts before running them
This may not apply to everyone, but the scripts I wrote to be executed
after Katello finishes installing are to be run in order. For the most
part they seemed to go in order, but I upgraded rubies and that seems to
be less consistent now. I added an explicit call to sort.

```
# directory structure
./scripts
  10-packages.bash
  20-dotfiles.bash
  30-foretello.bash
  40-hammer.bash
```