allow hammer remote install command to time out
Previously, this command would occasionally time out even after the
package was successfully installed. This causes bats to fail.

I put in a redmine issue for the erroneous timeout, see
http://projects.theforeman.org/issues/15089. This workaround should
allow bats to pass. Note that we still check for the package after the
command times out, and will fail the job if the package is not
installed.