Improve the hammer devel environment
This change removes rbenv in favor of an SCL installation, greatly
reducing the amount of time and CPU resources required to up this box.

In addition, it allows the user to use the `hammer` executable from any
directory once ssh'd into the box.

Finally, the etc_hosts role is added to this playbook to ensure
whichever katello/satellite server the user wishes to use as the
hammer host is already defined in /etc/hosts