Deploy smart proxy as a sidecar to Pulp
This works around a current design inconsistency in Katello where
it detects Pulp communication differently depending on code path.
This ensures that the Smart Proxy that has Pulp feature runs under
the same service reference as Pulp ensuring the hostname referenced
for communication is the same. This fixes a bug with some operations
such as Repository updates.

To be re-worked with the resolution of:

https://github.com/Katello/katello/pull/7470