Add luna_provisioning playbook
luna_provisioning is built on katello_provisioning.
It contains the same plugin configuration as plain luna,
but adds provisioning setup + seeded data for some of the plugins

I split the plugin roles into install and seed roles,
mainly because I aim to reuse the seed part in downstream.

Other changes:

* I added missing subscriptions to activation key

* Use example.com domain for provisioned hosts rather than
  the one on which foreman runs

* Add and sync plugins repo to have foreman_scap_client available
  on disconnected setups (demo hardware)