Drop Katello 3.8 and 3.9
These are no longer supported and shouldn't be shown anymore. Previously
we included these to test upgrades but due to a candlepin issue a fresh
installation is no longer possible without workarounds. On the Foreman
side the 3.10 release pipeline will drop upgrade testing. Note that 3.10
is out of support anyway, but it's still needed to push in critical
upgrade fixes.