Move foreman-initial-admin-password to install.yaml
Rather than keeping it as a default value, this moves it to install.yaml
which has more information about whether it really needs to be there.
Besides scenarios that are not foreman-proxy-content (which doesn't
include the foreman class), it also doesn't make sense on upgrades since
it's the initial password.

A nice benefit is that it isn't spread over the various playbooks
anymore.