Tune the Puppetserver in the devel scenario
This tunes the Puppetserver for a minimal deployment where very few to
no clients are expected. This keeps the memory usage of the devel box a
bit in check. By doing this via the default answers file, it shouldn't
affect users who've already specified other options on the command line.