minor changes to Ansible instructions
I don't think tying to a specific version of foreman and katello in the docs is a good idea, unless someone is going to update it for every version. Also, some changes to the local deployment. Obviously, this is not the place for full Ansible instructions, but if people are new to repos, the fact that 'yum install ansible' does not work without EPEL is an easy blocker to avoid.