Update production.md
Changed Ansible instructions based on PR feedback for both local deployment and remote deployment. Some of the steps between local and remote deployment are duplicated, which is not ideal for future edits, but makes the document more scanable. 

Inspected the Vagrant versioning and determined that mirroring vagrant status version output with the vagrant up command is obvious (I say this with very little vagrant experience, but with enough high level understanding of vagrant I feel qualified to make that call).