Add automatic disk resizing when disk_size is specified
If a box is defined with a larger disk size through the use of
disk_size parameter the user then has to manually resize the filesystem.
With the introduction of CentOS 8, which has a base disk size of
10GB within the image this operation will be prevalent for nearly
all Katello production and development deployments through Vagrant.

This adds automatic disk resizing by running a playbook as a
provisioner if the disk_size parameter is defined.