Identify proxy in bats test by ID 2
With Pulp 3, there is no feature that identifies a Pulp mirror
vs a Pulp server. This information is in the capabilities which are
not currently reported. Given the bats tests are predictable,
we can assume the external proxy ID will be 2 to handle both
Pulp 2 and Pulp 3 based tests cases.