Fix warning about string conversions
This fixes the following warning with sysctl:

    [WARNING]: The value 100000 (type int) in a string field was converted to u'100000' (type string). If this does not look like what you expect, quote the entire value to ensure it does not change.

This fixes the following warning with yum_repository:

    [WARNING]: The value 1 (type int) in a string field was converted to u'1' (type string). If this does not look like what you expect, quote the entire value to ensure it does not change.

Note that the type should be a string, but then the documentation states
it should be an integer from 1 to 99.

There was no warning about gpgcheck, but this follows valid values
according to the documentation. It's also used elsewhere in the same
file so it makes it more consistent.