gather facts about server before running smoker
smoker needs to be able to talk to the main server box, and we use
    hostvars[forklift_server_name].ansible_nodename
to get the FQDN of the server for configuring smoker to do so.
These facts are unknown to Ansible in this play as it does not directly
affect forklift_server_name -- setup module with delegete_facts solves
that for us.