Use pytest-xdist in smoker to launch 2 processes
In my experience, there is a speedup if you use multiple processes since
there are times you wait. The number 2 was chosen because a large number
of our tests runs a browser and that can take a bit of memory. Even on
my laptop I saw diminishing returns after a while. 2 is a fairly
conservative number, but can already help.