Enhance to Ansible production instructions
I don't think tying to a specific version of foreman and katello in the docs is a good idea, unless someone is going to update it for every version. Also, some changes to the local deployment. Obviously, this is not the place for full Ansible instructions, but if people are new to repos, the fact that 'yum install ansible' does not work without EPEL is an easy blocker to avoid.

Changed Ansible instructions based on PR feedback for both local deployment and remote deployment. Some of the steps between local and remote deployment are duplicated, which is not ideal for future edits, but makes the document more scanable. 

Inspected the Vagrant versioning and determined that mirroring vagrant status version output with the vagrant up command is obvious (I say this with very little vagrant experience, but with enough high level understanding of vagrant I feel qualified to make that call).

Co-authored-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>