Use foreman-installer-katello instead of meta packages
This stops installing katello/foreman-proxy-content meta packages. The
reason is that on EL8 you need to enable the pki-core module. In
517d8ddcb77f6fb97525cb4efd35db5212c301fb this was removed with the
expectation that the installer takes care of it. This makes the katello
meta-package uninstallable.

puppet-katello already ensures the katello meta package is installed.
For foreman-proxy-content this isn't done so an installer change is
needed if this package is actually needed.