Add upgrade step to work around rubygem-foreman_remote_execution_core breaking nightly
This works around https://projects.theforeman.org/issues/32897 temporarily.
On upgrade, rubygem-foreman_remote_execution_core is still present on
the system, which enabls the plugin and causes foreman-proxy to throw
errors due to missing keys. This creates an inconsistency as the plugin
is enabled on the system but not in the installer. This package
is being deprecated and will get obsoleted when the issue is fixed.