Allow setting ansible provisioner config file per box
Sometimes it is useful to pass a different Ansible configuration
file for some boxes only, and Vagrant allows us to do so. This
commit exposes that setting in the box definition for Forklift.
If not explicitly specified, the Ansible provisioner will follow
the current default behavior of allowing Ansible to choose its
configuration file via its normal order of precedence, selecting
the ansible.cfg file local to this repository.