use ansible_host not ansible_nodename to access foreman in pipelines
ansible_nodename is the "real" hostname (as in `hostname -f`), which we
update in some tests via `katello-change-hostname` but do not properly
propagate to poor-CIs-DNS (aka /etc/hosts).
instead of trying to fix DNS, we can just configure smoker to use
ansible_host, which is the value from the inventory and that is always
reachable (as it is the IP in the case of the Vagrant inventory).