don't set `priority: 1` when setting up staging repos
Setting `priority=1` means that DNF will always pick packages from this
repo over the ones available from other repos, even if those have a
higher version [1]:

     If there is more than one candidate package for a particular operation,
     the one from a repo with the lowest priority value is picked, possibly
     despite being less convenient otherwise (e.g. by being a lower version).

This breaks the usage of packit builds, as those don't have a priority
set by default (thus getting the default `99` from DNF).

This also means that we potentially test a package that after promotion
from staging to prod will never be picked up by the end-user's system as
we do not (and should not) set a priority in `foreman-release` (and
others).

[1] https://dnf.readthedocs.io/en/latest/conf_ref.html